feat:07/01, 북마크 추가

+++ backend/src/main/java/com/team2/backend/domain/reservation/TimelistRepository.java
@@ -10,5 +10,5 @@ public interface TimelistRepository extends JpaRepository<Timelist, Long> {
 
     Timelist findByCheckNo(Long checkNo);
     @Query("SELECT timeNo FROM Timelist WHERE checkNo = :checkNo")
-    Integer[] findAllByCheckNo(@Param("checkNo")Long checkNo);
+    Long[] findAllByCheckNo(@Param("checkNo")Long checkNo);
 }
